对IntField内的getValue函数返回类型更改做了适配

+++ src/main/java/db/file/BTree/BTreeHeaderPage.java
@@ -48,14 +48,14 @@ public class BTreeHeaderPage implements Page {
 		// Read the next and prev pointers
 		try {
 			Field f = Type.INT_TYPE.parse(dis);
-			this.nextPage = ((IntField) f).getValue();
+			this.nextPage = (Integer)((IntField) f).getValue();
 		} catch (java.text.ParseException e) {
 			e.printStackTrace();
 		}
 
 		try {
 			Field f = Type.INT_TYPE.parse(dis);
-			this.prevPage = ((IntField) f).getValue();
+			this.prevPage = (Integer) ((IntField) f).getValue();
 		} catch (java.text.ParseException e) {
 			e.printStackTrace();
 		}
